+++ docker-gen-proto.sh
@@ -39,7 +39,7 @@ if [ -z "$1" ]; then
   #gen_proto_ruby email
   #gen_proto_ts frontend
   #gen_proto_js payment
-  gen_proto_go productcatalogservice
+  gen_proto_go product-catalog
   #gen_proto_php quote
   gen_proto_python recommendation
   #gen_proto_rust shipping