+++ ide-gen-proto.sh
@@ -74,7 +74,7 @@ gen_proto_go checkoutservice
 gen_proto_ts frontend
 gen_proto_ts react-native-app
 gen_proto_js payment
-gen_proto_go productcatalogservice
+gen_proto_go product-catalog
 # gen_proto_php quote
 gen_proto_python recommendation
 gen_proto_rust shipping