+++ test/tracetesting/run.bash
@@ -8,7 +8,7 @@
 set -e
 
 # Availalble services to test
-ALL_SERVICES=("ad" "cart" "currency" "checkoutservice" "frontend" "email" "payment" "productcatalogservice" "recommendation" "shipping")
+ALL_SERVICES=("ad" "cart" "currency" "checkoutservice" "frontend" "email" "payment" "product-catalog" "recommendation" "shipping")
 
 ## Script variables
 # Will contain the list of services to test
@@ -48,8 +48,8 @@ spec:
       value: $FRONTEND_ADDR
     - key: PAYMENT_ADDR
       value: $PAYMENT_ADDR
-    - key: PRODUCT_CATALOG_SERVICE_ADDR
-      value: $PRODUCT_CATALOG_SERVICE_ADDR
+    - key: PRODUCT_CATALOG_ADDR
+      value: $PRODUCT_CATALOG_ADDR
     - key: RECOMMENDATION_ADDR
       value: $RECOMMENDATION_ADDR
     - key: SHIPPING_ADDR